+++ internal/controller/operator/factory/vmagent/probe_test.go
@@ -94,7 +94,102 @@ relabel_configs:
 `,
 	})
 
-	// with ingress discover
+	// with Service discovery
+	f(opts{
+		cr: &vmv1beta1.VMAgent{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "default-vmagent",
+				Namespace: "default",
+			},
+		},
+		sc: &vmv1beta1.VMProbe{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "probe-service",
+				Namespace: "monitor",
+			},
+			Spec: vmv1beta1.VMProbeSpec{
+				Module:       "http200",
+				VMProberSpec: vmv1beta1.VMProberSpec{URL: "blackbox:9115"},
+				Targets: vmv1beta1.VMProbeTargets{
+					K8s: []*vmv1beta1.VMProbeTargetK8s{
+						{
+							Role:              k8sSDRoleService,
+							NamespaceSelector: vmv1beta1.NamespaceSelector{},
+							RelabelConfigs: []*vmv1beta1.RelabelConfig{
+								{
+									SourceLabels: []string{"label1"},
+									TargetLabel:  "api",
+									Action:       "replacement",
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+		want: `job_name: probe/monitor/probe-service
+honor_labels: false
+metrics_path: /probe
+params:
+  module:
+  - http200
+kubernetes_sd_configs:
+- role: service
+  namespaces:
+    names:
+    - monitor
+relabel_configs:
+- action: replace
+  source_labels:
+  - __meta_kubernetes_service_annotation_operator_victoriametrics_com_probe_path
+  target_label: __meta_kubernetes_service_path
+- if: '{__meta_kubernetes_service_annotationpresent_operator_victoriametrics_com_probe_path!="true"}'
+  action: replace
+  target_label: __meta_kubernetes_service_path
+  replacement: /
+- action: replace
+  source_labels:
+  - __meta_kubernetes_service_annotation_operator_victoriametrics_com_probe_scheme
+  target_label: __meta_kubernetes_service_scheme
+- if: '{__meta_kubernetes_service_annotationpresent_operator_victoriametrics_com_probe_scheme!="true"}'
+  action: replace
+  target_label: __meta_kubernetes_service_scheme
+  replacement: http
+- source_labels:
+  - __address__
+  separator: ;
+  regex: (.*)
+  target_label: __tmp_service_address
+  replacement: $1
+  action: replace
+- source_labels:
+  - __meta_kubernetes_service_scheme
+  - __address__
+  - __meta_kubernetes_service_path
+  separator: ;
+  regex: (.+);(.+);(.+)
+  target_label: __param_target
+  replacement: ${1}://${2}${3}
+  action: replace
+- source_labels:
+  - __meta_kubernetes_namespace
+  target_label: namespace
+- source_labels:
+  - __meta_kubernetes_service_name
+  target_label: service
+- source_labels:
+  - label1
+  target_label: api
+  action: replacement
+- source_labels:
+  - __param_target
+  target_label: instance
+- target_label: __address__
+  replacement: blackbox:9115
+`,
+	})
+
+	// with Ingress discovery
 	f(opts{
 		cr: &vmv1beta1.VMAgent{
 			ObjectMeta: metav1.ObjectMeta{
@@ -111,13 +206,16 @@ relabel_configs:
 				Module:       "http200",
 				VMProberSpec: vmv1beta1.VMProberSpec{URL: "blackbox:9115"},
 				Targets: vmv1beta1.VMProbeTargets{
-					Ingress: &vmv1beta1.ProbeTargetIngress{
-						NamespaceSelector: vmv1beta1.NamespaceSelector{},
-						RelabelConfigs: []*vmv1beta1.RelabelConfig{
-							{
-								SourceLabels: []string{"label1"},
-								TargetLabel:  "api",
-								Action:       "replacement",
+					K8s: []*vmv1beta1.VMProbeTargetK8s{
+						{
+							Role:              k8sSDRoleIngress,
+							NamespaceSelector: vmv1beta1.NamespaceSelector{},
+							RelabelConfigs: []*vmv1beta1.RelabelConfig{
+								{
+									SourceLabels: []string{"label1"},
+									TargetLabel:  "api",
+									Action:       "replacement",
+								},
 							},
 						},
 					},
@@ -160,6 +258,142 @@ relabel_configs:
   target_label: ingress
 - source_labels:
   - label1
+  target_label: api
+  action: replacement
+- source_labels:
+  - __param_target
+  target_label: instance
+- target_label: __address__
+  replacement: blackbox:9115
+`,
+	})
+
+	// with mixed discovery
+	f(opts{
+		cr: &vmv1beta1.VMAgent{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "default-vmagent",
+				Namespace: "default",
+			},
+		},
+		sc: &vmv1beta1.VMProbe{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      "probe-mixed",
+				Namespace: "monitor",
+			},
+			Spec: vmv1beta1.VMProbeSpec{
+				Module:       "http200",
+				VMProberSpec: vmv1beta1.VMProberSpec{URL: "blackbox:9115"},
+				Targets: vmv1beta1.VMProbeTargets{
+					K8s: []*vmv1beta1.VMProbeTargetK8s{
+						{
+							Role:              k8sSDRoleIngress,
+							NamespaceSelector: vmv1beta1.NamespaceSelector{},
+							RelabelConfigs: []*vmv1beta1.RelabelConfig{
+								{
+									SourceLabels: []string{"label1"},
+									TargetLabel:  "api",
+									Action:       "replacement",
+								},
+							},
+						},
+						{
+							Role:              k8sSDRoleService,
+							NamespaceSelector: vmv1beta1.NamespaceSelector{},
+							RelabelConfigs: []*vmv1beta1.RelabelConfig{
+								{
+									SourceLabels: []string{"label2"},
+									TargetLabel:  "api",
+									Action:       "replacement",
+								},
+							},
+						},
+					},
+				},
+			},
+		},
+		want: `job_name: probe/monitor/probe-mixed
+honor_labels: false
+metrics_path: /probe
+params:
+  module:
+  - http200
+kubernetes_sd_configs:
+- role: ingress
+  namespaces:
+    names:
+    - monitor
+- role: service
+  namespaces:
+    names:
+    - monitor
+relabel_configs:
+- source_labels:
+  - __address__
+  separator: ;
+  regex: (.*)
+  target_label: __tmp_ingress_address
+  replacement: $1
+  action: replace
+- source_labels:
+  - __meta_kubernetes_ingress_scheme
+  - __address__
+  - __meta_kubernetes_ingress_path
+  separator: ;
+  regex: (.+);(.+);(.+)
+  target_label: __param_target
+  replacement: ${1}://${2}${3}
+  action: replace
+- source_labels:
+  - __meta_kubernetes_namespace
+  target_label: namespace
+- source_labels:
+  - __meta_kubernetes_ingress_name
+  target_label: ingress
+- source_labels:
+  - label1
+  target_label: api
+  action: replacement
+- action: replace
+  source_labels:
+  - __meta_kubernetes_service_annotation_operator_victoriametrics_com_probe_path
+  target_label: __meta_kubernetes_service_path
+- if: '{__meta_kubernetes_service_annotationpresent_operator_victoriametrics_com_probe_path!="true"}'
+  action: replace
+  target_label: __meta_kubernetes_service_path
+  replacement: /
+- action: replace
+  source_labels:
+  - __meta_kubernetes_service_annotation_operator_victoriametrics_com_probe_scheme
+  target_label: __meta_kubernetes_service_scheme
+- if: '{__meta_kubernetes_service_annotationpresent_operator_victoriametrics_com_probe_scheme!="true"}'
+  action: replace
+  target_label: __meta_kubernetes_service_scheme
+  replacement: http
+- source_labels:
+  - __address__
+  separator: ;
+  regex: (.*)
+  target_label: __tmp_service_address
+  replacement: $1
+  action: replace
+- source_labels:
+  - __meta_kubernetes_service_scheme
+  - __address__
+  - __meta_kubernetes_service_path
+  separator: ;
+  regex: (.+);(.+);(.+)
+  target_label: __param_target
+  replacement: ${1}://${2}${3}
+  action: replace
+- source_labels:
+  - __meta_kubernetes_namespace
+  target_label: namespace
+- source_labels:
+  - __meta_kubernetes_service_name
+  target_label: service
+- source_labels:
+  - label2
   target_label: api
   action: replacement
 - source_labels:
@@ -313,16 +547,19 @@ basic_auth:
 				Module:       "http200",
 				VMProberSpec: vmv1beta1.VMProberSpec{URL: "blackbox:9115"},
 				Targets: vmv1beta1.VMProbeTargets{
-					Ingress: &vmv1beta1.ProbeTargetIngress{
-						NamespaceSelector: vmv1beta1.NamespaceSelector{},
-						Selector: *metav1.SetAsLabelSelector(map[string]string{
-							"ingress-class": "ngin",
-						}),
-						RelabelConfigs: []*vmv1beta1.RelabelConfig{
-							{
-								SourceLabels: []string{"label1"},
-								TargetLabel:  "api",
-								Action:       "replacement",
+					K8s: []*vmv1beta1.VMProbeTargetK8s{
+						{
+							Role:              k8sSDRoleIngress,
+							NamespaceSelector: vmv1beta1.NamespaceSelector{},
+							Selector: *metav1.SetAsLabelSelector(map[string]string{
+								"ingress-class": "nginx",
+							}),
+							RelabelConfigs: []*vmv1beta1.RelabelConfig{
+								{
+									SourceLabels: []string{"label1"},
+									TargetLabel:  "api",
+									Action:       "replacement",
+								},
 							},
 						},
 					},
@@ -342,7 +579,7 @@ kubernetes_sd_configs:
     - monitor
   selectors:
   - role: ingress
-    label: ingress-class=ngin
+    label: ingress-class=nginx
 relabel_configs:
 - source_labels:
   - __address__